roken: Fix size type in resolve.c

+++ lib/roken/resolve.c
@@ -628,7 +628,7 @@ rk_dns_srv_order(struct rk_dns_reply *r)
 {
     struct rk_resource_record **srvs, **ss, **headp;
     struct rk_resource_record *rr;
-    int num_srv = 0;
+    size_t num_srv = 0;
 
     rk_random_init();
